Keycard Door and Switch Door
Keycard and also Keycard door are made just make sure the keycard door are the same variable or it wont work. With the switch door and switch when you pick the door hit the eyedrop and pick the switch you want it to link to or it also wont work. Each of these will interact with the door by pressing e if you set them up correctly. Working on a weight switch later.